Update CheckBlocks start
 - Fix getBlock parameters
 - Trace first blocks errors
 - Skip check transactions at start

+++ src/services/classes/CheckBlocks.js
@@ -12,13 +12,18 @@ export class CheckBlocks extends BlocksBase {
     this.tipSize = options.bcTipSize || 12
   }
 
+  async saveTipBlocks () {
+    await this.getBlock(0).catch(err => this.log.trace(err))
+    let lastBlock = await this.getLastBlock().catch(err => this.log.trace(err))
+    await this.getBlock(lastBlock.hash).catch(err => this.log.trace(err))
+  }
   async start (emitter) {
     try {
       if (emitter) this.setEmitter(emitter)
       if (!this.emit) throw new Error('The emitter should be defined')
-      await Promise.all([this.getBlock(0), this.getLastBlock()])
+      await this.saveTipBlocks()
       this.log.info('Checking database')
-      let res = await this.checkDb(true)
+      let res = await this.checkDb({ checkOrphans: true, skipTxs: true })
       this.log.info('Getting missing blocks')
       if (!res) this.log.info('There are no missing blocks')
       else this.log.trace(res)
@@ -29,7 +34,7 @@ export class CheckBlocks extends BlocksBase {
     }
   }
 
-  async checkDb (checkOrphans, lastBlock, firstBlock) {
+  async checkDb ({ checkOrphans, lastBlock, firstBlock, skipTxs }) {
     if (!lastBlock || !lastBlock.number) lastBlock = await this.getHighDbBlock()
     if (!lastBlock) return
     lastBlock = lastBlock.number
@@ -39,9 +44,11 @@ export class CheckBlocks extends BlocksBase {
     if (blocks < lastBlock + 1) {
       missingSegments = await this.getMissingSegments()
     }
-
-    let missingTxs = await this.getMissingTransactions(lastBlock, firstBlock)
-    await this.deleteMissingTxsBlocks(missingTxs)
+    let missingTxs
+    if (!skipTxs) {
+      missingTxs = await this.getMissingTransactions(lastBlock, firstBlock)
+      await this.deleteMissingTxsBlocks(missingTxs)
+    }
 
     let res = { lastBlock, blocks, missingSegments, missingTxs }
     if (checkOrphans) {
@@ -96,8 +103,8 @@ export class CheckBlocks extends BlocksBase {
   }
 
   async getBlock (hashOrNumber) {
-    const { nod3, collections, log, nativeContracts } = this
-    return getBlock(hashOrNumber, { nod3, collections, log, nativeContracts })
+    const { nod3, collections, log, initConfig } = this
+    return getBlock(hashOrNumber, { nod3, collections, log, initConfig })
   }
 
   getBlockFromDb (hashOrNumber) {
@@ -177,7 +184,8 @@ export class CheckBlocks extends BlocksBase {
         let lastBlock = this.tipBlock
         this.tipCount = 0
         this.log.info(`Checking db / LastBlock: ${lastBlock}`)
-        let res = await this.checkDb(true, lastBlock, lastBlock - this.tipSize * 10)
+        let firstBlock = lastBlock - this.tipSize * 10
+        let res = await this.checkDb({ checkOrphans: true, lastBlock, firstBlock })
         this.log.trace(`Check db: ${res}`)
         return this.getBlocks(res)
       }
